refactored folder layout
