+++ public/components/featured-row.tsx
