+++ public/components/featured.module.css
