+++ public/components/featured.tsx
@@ -1,8 +1,8 @@
 import type { FunctionalComponent } from 'preact'
-import componentsUrl from 'url:../../images/svg/components.svg'
-import computerUrl from 'url:../../images/svg/computer.svg'
-import iaUrl from 'url:../../images/svg/ia.svg'
-import sourceWormUrl from 'url:../../images/svg/source-worm.svg'
+import componentsUrl from 'url:../images/svg/components.svg'
+import computerUrl from 'url:../images/svg/computer.svg'
+import iaUrl from 'url:../images/svg/ia.svg'
+import sourceWormUrl from 'url:../images/svg/source-worm.svg'
 import { FeaturedRow } from './featured-row'
 import styles from './featured.module.css'
