+++ public/components/index.tsx
@@ -2,9 +2,9 @@ import type { FunctionalComponent } from 'preact'
 import { Layout } from '../components/layout'
 import { categories, categoryToSlugs } from '../resources'
 import styles from './index.module.css'
-import { Featured } from './index/featured'
-import { GettingStartedAndNav } from './index/gs-and-nav'
-import { Header } from './index/header'
+import { Featured } from './featured'
+import { GettingStartedAndNav } from './gs-and-nav'
+import { Header } from './header'
 
 export const Index: FunctionalComponent = () => {
   return (
