+++ public/components/resources-list.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useRef } from 'preact/hooks'
 import { useDB } from '../db-context'
 import { useFetched } from '../hooks/use-fetched'
 import { isPrerenderContext } from '../prerender-context'
-import  arrowURL from 'url:..//images/svg/arrow.svg'
+import  arrowURL from 'url:../images/svg/arrow.svg'
 import type { Resource } from '../resources'
 import { categoryToSlugs } from '../resources'
 import styles from './resources-list.module.css'
